--- a/velox/core/CoreTypeSystem.h
+++ b/velox/core/CoreTypeSystem.h
@@ -241,7 +241,7 @@ struct IMapVal {
   static std::shared_ptr<const Type> veloxType() {
     return MAP(UdfToType<KEY>::veloxType(), UdfToType<VAL>::veloxType());
   }
-  using container_t = typename std::unordered_map<KEY, std::optional<VAL>>;
+  using container_t = typename folly::F14FastMap<KEY, std::optional<VAL>>;
   using iterator = typename container_t::iterator;
   using reference = typename container_t::reference;
   using const_iterator = typename container_t::const_iterator;