--- a/velox/functions/prestosql/benchmarks/MapInputBenchmark.cpp
+++ b/velox/functions/prestosql/benchmarks/MapInputBenchmark.cpp
@@ -1,0 +1,175 @@
+/*
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+#include <folly/Benchmark.h>
+#include <folly/init/Init.h>
+#include "velox/functions/Macros.h"
+#include "velox/functions/lib/benchmarks/FunctionBenchmarkBase.h"
+#include "velox/functions/prestosql/VectorFunctions.h"
+
+using namespace facebook::velox;
+using namespace facebook::velox::exec;
+
+namespace {
+
+// Vector function implementation
+class MapSumValuesAndKeysVector : public exec::VectorFunction {
+ public:
+  void apply(
+      const SelectivityVector& rows,
+      std::vector<VectorPtr>& args,
+      const TypePtr& /* outputType */,
+      exec::EvalCtx* context,
+      VectorPtr* result) const override {
+    auto arg = args.at(0);
+
+    auto mapVector = arg->as<MapVector>();
+    exec::LocalDecodedVector mapHolder(context, *arg, rows);
+    auto mapIndices = mapHolder->indices();
+
+    auto mapKeys = mapVector->mapKeys();
+    exec::LocalDecodedVector mapKeysHolder(context, *mapKeys, rows);
+    auto decodedMapKeys = mapKeysHolder.get();
+
+    auto mapValues = mapVector->mapValues();
+    exec::LocalDecodedVector mapValuesHolder(context, *mapValues, rows);
+    auto decodedMapValues = mapKeysHolder.get();
+
+    auto rawOffsets = mapVector->rawOffsets();
+    auto rawSizes = mapVector->rawSizes();
+
+    BaseVector::ensureWritable(rows, BIGINT(), context->pool(), result);
+    auto flatResult = (*result)->asFlatVector<int64_t>();
+
+    rows.applyToSelected([&](vector_size_t row) {
+      size_t mapIndex = mapIndices[row];
+      size_t offsetStart = rawOffsets[mapIndex];
+      size_t offsetEnd = offsetStart + rawSizes[mapIndex];
+
+      auto sum = 0;
+      for (size_t offset = offsetStart; offset < offsetEnd; ++offset) {
+        sum += decodedMapKeys->valueAt<int64_t>(offset);
+        sum += decodedMapValues->valueAt<int64_t>(offset);
+      }
+      flatResult->set(row, sum);
+    });
+  }
+};
+
+// Simple function implementation
+template <typename T>
+struct MapSumValuesAndKeysSimple {
+  VELOX_DEFINE_FUNCTION_TYPES(T);
+
+  FOLLY_ALWAYS_INLINE bool call(
+      int64_t& out,
+      const arg_type<Map<int64_t, int64_t>>& map) {
+    out = 0;
+    for (const auto& entry : map) {
+      out += entry.first;
+      out += *entry.second;
+    }
+    return true;
+  }
+};
+
+class MapInputBenchmark : public functions::test::FunctionBenchmarkBase {
+ public:
+  MapInputBenchmark() : FunctionBenchmarkBase() {
+    registerFunction<MapSumValuesAndKeysSimple, int64_t, Map<int64_t, int64_t>>(
+        {"map_sum_simple"});
+    facebook::velox::exec::registerVectorFunction(
+        "map_sum_vector",
+        {exec::FunctionSignatureBuilder()
+             .returnType("bigint")
+             .argumentType("map(bigint,bigint)")
+             .variableArity()
+             .build()},
+        std::make_unique<MapSumValuesAndKeysVector>());
+  }
+
+  RowVectorPtr makeData() {
+    const vector_size_t size = 1'000;
+    auto arrayVector = vectorMaker_.mapVector<int64_t, int64_t>(
+        size,
+        [](auto row) { return row % 100; },
+        [](auto row) { return row % 100; },
+        [](auto row) { return row % 100; });
+
+    return vectorMaker_.rowVector({arrayVector});
+  }
+
+  void run(const std::string& functionName) {
+    folly::BenchmarkSuspender suspender;
+    auto rowVector = makeData();
+    auto exprSet = compileExpression(
+        fmt::format("{}(c0)", functionName), rowVector->type());
+    suspender.dismiss();
+
+    doRun(exprSet, rowVector);
+  }
+
+  void doRun(ExprSet& exprSet, const RowVectorPtr& rowVector) {
+    int cnt = 0;
+    for (auto i = 0; i < 100; i++) {
+      cnt += evaluate(exprSet, rowVector)->size();
+    }
+    folly::doNotOptimizeAway(cnt);
+  }
+
+  bool test() {
+    auto rowVector = makeData();
+    auto exprSet1 =
+        compileExpression(fmt::format("map_sum_vector(c0)"), rowVector->type());
+    auto exprSet2 =
+        compileExpression(fmt::format("map_sum_simple(c0)"), rowVector->type());
+    auto result1 = evaluate(exprSet1, rowVector);
+    auto result2 = evaluate(exprSet2, rowVector);
+
+    auto flatResults1 = result1->asFlatVector<int64_t>();
+    auto flatResults2 = result2->asFlatVector<int64_t>();
+
+    if (flatResults1->size() != flatResults2->size()) {
+      return false;
+    }
+
+    for (auto i = 0; i < flatResults1->size(); i++) {
+      if (flatResults1->valueAt(i) != flatResults2->valueAt(i)) {
+        return false;
+      }
+    }
+
+    return true;
+  }
+};
+
+BENCHMARK(vectorFunction) {
+  MapInputBenchmark benchmark;
+  benchmark.run("map_sum_vector");
+}
+
+BENCHMARK_RELATIVE(simpleFunction) {
+  MapInputBenchmark benchmark;
+  benchmark.run("map_sum_simple");
+}
+} // namespace
+
+int main(int /*argc*/, char** /*argv*/) {
+  MapInputBenchmark benchmark;
+  if (benchmark.test()) {
+    folly::runBenchmarks();
+  }
+  return 0;
+}