initial mapView implementation.
numbers for the benchmarks are better but there is still a significant gap!
3X-4X 
will try to optimize it a bit more.
```
============================================================================
mapSumVectorFunction                                               2.81ms   356.33
mapSumSimpleFunction                                    38.48%     7.29ms   137.11
nestedMapSumVectorFunction                                      5.98ms   167.15
nestedMapSumSimpleFunction                           25.83%    23.16ms    43.17
============================================================================
```
- this need to be rebased on top of arrayView and the benchmark diffs and reorg shared components.
- iterators for values only and keys only can be created.

In order to understand the 3x difference i added a benchmark where a vector reader is defined inside a vector function 
and the whole function is written using the simple function mapView reader. (see the new nested_map_sum_vector_mapview in the diff) . results are the same

```
============================================================================
nestedSumVectorFunction                                      7.47ms   133.87
nestedSumSimpleFunction                           29.49%    25.33ms    39.48
nestedSumVectorFunctionMapView                    34.26%    21.80ms    45.87
============================================================================
```
next i will compare this mapView with @mbasmanova  initial solution to see if its faster.

I tried the prev approach of masha and its same numbers:
```
============================================================================
vectorFunction                                               3.69ms   271.15
simpleFunction                                    38.49%     9.58ms   104.37
============================================================================
```

my suspicion is that the usage of decodedResults Vs decoder. and the way the index() function 
is computed in decoder is more efficient, in addition to the copying done for the decodedResults
allocation/deallocation. could all be the reason of the slowdown
I will investigate that next.'

update: by updating the decodedResults to avoid un-needed materialization of indicies() we get 2x more speedup and the current results are as the following:
```
============================================================================
vectorFunction                                               2.87ms   347.98
simpleFunction                                    84.87%     3.39ms   295.32
nestedSumVectorFunction                                      5.95ms   168.09
nestedSumSimpleFunction                           54.20%    10.98ms    91.11
nestedSumVectorFunctionMapView                    86.87%     6.85ms   146.02
============================================================================
```